ai/requirements.txt
fastapi
uvicorn
python-multipart

pdfplumber
python-pptx

chromadb
openai

pydantic
python-dotenv
